refactor: relocate formating utils
