+++ src/mutmut/utils/format_utils.py
@@ -3,20 +3,17 @@
 import os
 from pathlib import Path
 
-from mutmut.mutation.trampoline_templates import CLASS_NAME_SEPARATOR
+CLASS_NAME_SEPARATOR = "ǁ"
 
 
-def make_mutant_key(func_name: str, class_name: str | None = None) -> str:
-    """Create a consistent key for identifying a function/method for mutation tracking.
-
-    :param func_name: The function or method name
-    :param class_name: The containing class name, or None for top-level functions
-    :return: A key string like "xǁMyClassǁmethod" for methods or "x_foo" for functions
-    """
+def mangle_function_name(*, name: str, class_name: str | None) -> str:
+    assert CLASS_NAME_SEPARATOR not in name
     if class_name:
-        return f"x{CLASS_NAME_SEPARATOR}{class_name}{CLASS_NAME_SEPARATOR}{func_name}"
+        assert CLASS_NAME_SEPARATOR not in class_name
+        prefix = f"x{CLASS_NAME_SEPARATOR}{class_name}{CLASS_NAME_SEPARATOR}"
     else:
-        return f"x_{func_name}"
+        prefix = "x_"
+    return f"{prefix}{name}"
 
 
 def parse_mutant_key(key: str) -> tuple[str, str | None]:
@@ -53,3 +50,34 @@ def get_mutant_name(relative_source_path: Path, mutant_method_name: str) -> str:
     mutant_name = f"{module_name}.{mutant_method_name}"
     mutant_name = mutant_name.replace(".__init__.", ".")
     return mutant_name
+
+
+def get_module_from_key(key: str) -> str:
+    """Extract module name from a mangled function key like 'app.foo.x_bar'.
+
+    The function name starts with 'x_' or 'xǁ', so we find that part
+    and return everything before it as the module path.
+    """
+    parts = key.split(".")
+    for i in range(len(parts) - 1, -1, -1):
+        if parts[i].startswith("x_") or parts[i].startswith("x\u01c1"):
+            return ".".join(parts[:i])
+    return key.rsplit(".", 1)[0] if "." in key else key
+
+
+def mangled_name_from_mutant_name(mutant_name: str) -> str:
+    assert "__mutmut_" in mutant_name, mutant_name
+    return mutant_name.partition("__mutmut_")[0]
+
+
+def orig_function_and_class_names_from_key(mutant_name: str) -> tuple[str, str | None]:
+    r = mangled_name_from_mutant_name(mutant_name)
+    _, _, r = r.rpartition(".")
+    class_name = None
+    if CLASS_NAME_SEPARATOR in r:
+        class_name = r[r.index(CLASS_NAME_SEPARATOR) + 1 : r.rindex(CLASS_NAME_SEPARATOR)]
+        r = r[r.rindex(CLASS_NAME_SEPARATOR) + 1 :]
+    else:
+        assert r.startswith("x_"), r
+        r = r[2:]
+    return r, class_name
